pyproject.toml
[project]
name = "content-monitoring"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "atproto>=0.0.65",
    "beartype>=0.22.9",
    "helium>=5.1.2",
    "ollama>=0.6.1",
    "pandas>=2.3.3",
    "polars>=1.36.1",
    "psycopg2>=2.9.11",
    "python-dotenv>=1.2.1",
    "spacy>=3.8.11",
]

[project.scripts]
scraping_powerbi = "src.__main__:main"
 

# -------------------------------
# Configuration Ruff
# -------------------------------
[tool.ruff]
target-version = "py313"
fix = true
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.pytest.ini_options]

testpaths = ["tests", "src"]

addopts = [
    "--doctest-modules",
    "--cov=src/scraping_powerbi",
    "--cov-report=term-missing",
    "--cov-report=html",
]

pythonpath = ["."]

[tool.coverage.run]
source = ["src"]
branch = true

[tool.coverage.report]
fail_under = 80
show_missing = true
exclude_lines = [
    "if __name__ == .__main__.:",
    "pragma: no cover",
    "raise NotImplementedError",
]
skip_empty = true

 
[tool.mypy]
python_version = "3.13"
strict = true

[dependency-groups]
dev = [
    "ruff>=0.14.9",
]
